# perf(loader): use a quicker version of vim.fs.normalize