@lewis6991 2024-05-15
Refined this some more and added some benchmarking data. Also added some optimisations to `path_resolve_dot()`.